anachnu po beyahad

+++ Ex03.ConsoleUI/UserInterface.cs
@@ -8,8 +8,8 @@ namespace Ex03.ConsoleUI
     internal abstract class UserInterface
     {
         protected Garage m_Garage;
-        protected static readonly Dictionary<byte, KeyValuePair<string, string>> sr_AvailableActionsForUser = 
-            new Dictionary<byte, KeyValuePair<string, string>>();
+        protected static readonly Dictionary<ushort, KeyValuePair<string, string>> sr_AvailableActionsForUser = 
+            new Dictionary<ushort, KeyValuePair<string, string>>();
 
         static UserInterface()
         {
